+++ examples/write-all.js
@@ -1,17 +1,17 @@
 'use strict'
 
-const PromiseWritable = require('../lib/promise-writable')
-const fs = require('fs')
+const { PromiseWritable } = require('../lib/promise-writable')
+const { createWriteStream } = require('fs')
 
 async function main () {
-  const wstream = new PromiseWritable(fs.createWriteStream(process.argv[2] || 'a.out'))
-  const size = Number(process.argv[3] || 1000)
-  const chunkSize = Number(process.argv[4] || 64 * 1024)
+  const wstream = new PromiseWritable(createWriteStream(process.argv[2] || 'a.out'))
+  const size = Number(process.argv[3]) || 1000
+  const chunkSize = Number(process.argv[4]) || 64 * 1024
 
   const content = Buffer.alloc(size)
 
-  await wstream.writeAll(content, chunkSize)
-  console.log(`Write ${size} bytes in total`)
+  const written = await wstream.writeAll(content, chunkSize)
+  console.info(`Written ${written} bytes in total`)
 }
 
-main()
+main().catch(console.error)