+++ test/promise-writable.js
@@ -15,17 +15,23 @@ Feature('Test promise-writable module', () => {
   class MockStream extends EventEmitter {
     constructor () {
       super()
+      this.bytesWritten = 0
       this.writable = true
       this._buffer = Buffer.alloc(0)
+      this.once('finish', () => { this.closed = true })
     }
     write (chunk) {
+      if (this.closed) { return this.emit('error', new Error('writeAll after end')) }
       this._buffer = Buffer.concat([this._buffer, chunk])
+      this.bytesWritten = this._buffer.length
       return !chunk.toString().startsWith('pause')
     }
+    close () { this.closed = true }
+    destroy () { this.destroyed = true }
     end () { }
   }
 
-  Scenario('Write chunks to stream which doesn not pause', () => {
+  Scenario('Write chunks to stream which does not pause', () => {
     let promise
     let promiseWritable
     let stream
@@ -43,7 +49,7 @@ Feature('Test promise-writable module', () => {
     })
 
     Then('promise is fulfilled', () => {
-      return promise.should.be.fulfilled.and.ok
+      return promise.should.eventually.equal(6)
     })
 
     And('stream should contain this chunk', () => {
@@ -55,7 +61,7 @@ Feature('Test promise-writable module', () => {
     })
 
     Then('promise is fulfilled', () => {
-      return promise.should.be.fulfilled.and.ok
+      return promise.should.eventually.equal(6)
     })
 
     And('stream should contain another chunk', () => {
@@ -85,7 +91,7 @@ Feature('Test promise-writable module', () => {
     })
 
     Then('promise is fulfilled', () => {
-      return promise.should.be.fulfilled.and.ok
+      return promise.should.eventually.equal(6)
     })
 
     And('stream should contain this chunk', () => {
@@ -101,7 +107,7 @@ Feature('Test promise-writable module', () => {
     })
 
     Then('promise is fulfilled', () => {
-      return promise.should.be.fulfilled.and.ok
+      return promise.should.eventually.equal(6)
     })
 
     And('stream should contain another chunk', () => {
@@ -110,7 +116,7 @@ Feature('Test promise-writable module', () => {
   })
 
   Scenario('Write chunk to already finished stream', () => {
-    let promise
+    let error
     let promiseWritable
     let stream
 
@@ -123,7 +129,7 @@ Feature('Test promise-writable module', () => {
     })
 
     When('I call write method', () => {
-      promiseWritable.write(Buffer.from('pause1'))
+      return promiseWritable.write(Buffer.from('chunk1'))
     })
 
     And('finish event is emitted', () => {
@@ -131,15 +137,77 @@ Feature('Test promise-writable module', () => {
     })
 
     And('I call write method again', () => {
-      promise = promiseWritable.write(Buffer.from('pause2'))
-    })
-
-    And('error event is emitted', () => {
-      stream.emit('error', new Error('write after end'))
+      return promiseWritable.write(Buffer.from('chunk2')).catch((err) => {
+        error = err
+      })
     })
 
     Then('promise is rejected', () => {
-      return promise.should.be.rejectedWith(Error, 'write after end')
+      return error.should.be.an('error', 'write after end')
+    })
+  })
+
+  Scenario('Write chunk to already closed stream', () => {
+    let error
+    let promiseWritable
+    let stream
+
+    Given('Writable object', () => {
+      stream = new MockStream()
+    })
+
+    And('PromiseWritable object', () => {
+      promiseWritable = new PromiseWritable(stream)
+    })
+
+    When('I call write method', () => {
+      return promiseWritable.write(Buffer.from('chunk1'))
+    })
+
+    And('stream is closed', () => {
+      stream.close()
+    })
+
+    And('I call write method again', () => {
+      return promiseWritable.write(Buffer.from('chunk2')).catch((err) => {
+        error = err
+      })
+    })
+
+    Then('promise is rejected', () => {
+      return error.should.be.an('error', 'write after end')
+    })
+  })
+
+  Scenario('Write chunk to already destroyed stream', () => {
+    let error
+    let promiseWritable
+    let stream
+
+    Given('Writable object', () => {
+      stream = new MockStream()
+    })
+
+    And('PromiseWritable object', () => {
+      promiseWritable = new PromiseWritable(stream)
+    })
+
+    When('I call write method', () => {
+      return promiseWritable.write(Buffer.from('chunk1'))
+    })
+
+    And('stream is destroyed', () => {
+      stream.destroy()
+    })
+
+    And('I call write method again', () => {
+      return promiseWritable.write(Buffer.from('chunk2')).catch((err) => {
+        error = err
+      })
+    })
+
+    Then('promise is rejected', () => {
+      return error.should.be.an('error', 'write after end')
     })
   })
 
@@ -191,7 +259,7 @@ Feature('Test promise-writable module', () => {
     })
 
     Then('promise is fulfilled', () => {
-      return promise.should.be.fulfilled.and.ok
+      return promise.should.eventually.equal(18)
     })
 
     And('stream should contain this chunk', () => {
@@ -221,7 +289,7 @@ Feature('Test promise-writable module', () => {
     })
 
     Then('promise is fulfilled', () => {
-      return promise.should.be.fulfilled.and.ok
+      return promise.should.eventually.equal(18)
     })
 
     And('stream should contain this chunk', () => {
@@ -257,7 +325,7 @@ Feature('Test promise-writable module', () => {
     })
 
     Then('promise is fulfilled', () => {
-      return promise.should.be.fulfilled.and.ok
+      return promise.should.eventually.equal(18)
     })
 
     And('stream should contain this chunk', () => {
@@ -282,8 +350,60 @@ Feature('Test promise-writable module', () => {
       promiseWritable.end()
     })
 
-    When('finish event is emitted', () => {
+    And('finish event is emitted', () => {
       stream.emit('finish')
+    })
+
+    And('I call writeAll method', () => {
+      promise = promiseWritable.writeAll(Buffer.from('pause1pause2pause3'))
+    })
+
+    Then('promise is rejected', () => {
+      return promise.should.be.rejectedWith(Error, 'writeAll after end')
+    })
+  })
+
+  Scenario('Write all to closed stream', () => {
+    let promise
+    let promiseWritable
+    let stream
+
+    Given('Writable object', () => {
+      stream = new MockStream()
+    })
+
+    And('PromiseWritable object', () => {
+      promiseWritable = new PromiseWritable(stream)
+    })
+
+    When('stream is closed', () => {
+      stream.close()
+    })
+
+    And('I call writeAll method', () => {
+      promise = promiseWritable.writeAll(Buffer.from('pause1pause2pause3'))
+    })
+
+    Then('promise is rejected', () => {
+      return promise.should.be.rejectedWith(Error, 'writeAll after end')
+    })
+  })
+
+  Scenario('Write all to destroyed stream', () => {
+    let promise
+    let promiseWritable
+    let stream
+
+    Given('Writable object', () => {
+      stream = new MockStream()
+    })
+
+    And('PromiseWritable object', () => {
+      promiseWritable = new PromiseWritable(stream)
+    })
+
+    When('stream is destroyed', () => {
+      stream.destroy()
     })
 
     And('I call writeAll method', () => {
@@ -348,8 +468,8 @@ Feature('Test promise-writable module', () => {
           return promise.should.eventually.equal('result')
         })
       } else {
-        Then('promise returns null', () => {
-          return promise.should.eventually.be.null
+        Then('promise returns undefined', () => {
+          return promise.should.eventually.be.undefined
         })
       }
     })
@@ -375,8 +495,8 @@ Feature('Test promise-writable module', () => {
         stream.emit('finish')
       })
 
-      Then('promise returns null', () => {
-        return promise.should.eventually.be.null
+      Then('promise returns undefined', () => {
+        return promise.should.eventually.be.undefined
       })
 
       When(`I call ${event} method`, () => {
@@ -389,7 +509,7 @@ Feature('Test promise-writable module', () => {
         })
       } else {
         Then('promise is fulfilled', () => {
-          return promise.should.eventually.be.null
+          return promise.should.eventually.be.undefined
         })
       }
     })
@@ -442,8 +562,8 @@ Feature('Test promise-writable module', () => {
       stream.emit('finish')
     })
 
-    Then('promise returns null', () => {
-      return promise.should.eventually.be.null
+    Then('promise returns undefined', () => {
+      return promise.should.eventually.be.undefined
     })
   })